% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_two.R
\name{add_two}
\alias{add_two}
\title{Add two numbers}
\usage{
add_two(a, b)
}
\arguments{
\item{a}{numeric input a}

\item{b}{numeric input b}
}
\value{
Numeric
}
\description{
Add two numbers
}
\details{
Adds two numeric inputs together. Numbers provided as strings
will be coerced to numeric.
}
\examples{
add_two(4, 5)
add_two("4", 10)
}
